+++ NEWS.md
@@ -18,10 +18,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Power off logo adapt to screen mode 
   [#101](https://github.com/orontee/taranis/issues/101)
 
+### Removed
+
+## [1.10.1 - 2024-11-23]
+
+### Changed
+
 - Fix wrong computation of alert description height
   [#100](https://github.com/orontee/taranis/issues/100)
-
-### Removed
 
 ## [1.10.0 - 2024-11-20]